A64.cpp
The program uses overloading of functions where if the function acts differently depending on the number or type of inputs. Here the printPrimeList will have to take two inputs in order for the program to find the prime numbers. However, the program can fill in the inputs with the overloaded getPrimeList functions. These three functions allows the user to have a functioning program even if they input 0, 1, or 2 inputs. The overloaded functions in this program provide default values to run the code.